backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-environ>=0.10.0
psycopg2-binary>=2.9.6
djangorestframework-simplejwt>=5.2.2
django-cors-headers>=4.0.0
drf-spectacular>=0.26.2
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
whitenoise>=6.4.0
gunicorn>=20.1.0
